+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 Release TBD
 - Fix layout for Agenda view
 - Allow different displayblock settings for list vs. calendar views.
+- Add `DateFunc::cal_days_in_month` in case PHP doesn't have calendar support.
 
 ## 1.5.7
 Release 2022-07-30